Move midi functionality out of lufa.c
