+++ tmk_core/common/host_driver.h
@@ -30,11 +30,6 @@ typedef struct {
     void (*send_mouse)(report_mouse_t *);
     void (*send_system)(uint16_t);
     void (*send_consumer)(uint16_t);
-#ifdef MIDI_ENABLE
-    void (*usb_send_func)(MidiDevice *, uint16_t, uint8_t, uint8_t, uint8_t);
-    void (*usb_get_midi)(MidiDevice *);
-    void (*midi_usb_init)(MidiDevice *);
-#endif
 } host_driver_t;
 
 #endif
