+++ tmk_core/protocol/midi.mk
@@ -5,6 +5,7 @@ SRC += midi.c \
 	   bytequeue/bytequeue.c \
 	   bytequeue/interrupt_setting.c \
 	   sysex_tools.c \
+     qmk_midi.c \
 	   $(LUFA_SRC_USBCLASS)
 
 VPATH += $(TMK_PATH)/$(MIDI_DIR)
