+++ tmk_core/protocol/midi/qmk_midi.h
@@ -1,4 +1,9 @@
 #pragma once
+
 #ifdef MIDI_ENABLE
-  MidiDevice midi_device;
+  #include "midi.h"
+  extern MidiDevice midi_device;
+  void setup_midi(void);
+  void send_midi_packet(MIDI_EventPacket_t* event);
+  bool recv_midi_packet(MIDI_EventPacket_t* const event);
 #endif
